refactor (WinningLotto): 값을 가지고 있는 객체에서 결과를 반환하도록 변경
