+++ src/main/kotlin/domain/LottoResult.kt
@@ -11,11 +11,5 @@ data class LottoResult(private val result: Map<Rank, Int>) : Map<Rank, Int> by r
 
     companion object {
         private const val ERROR_COUNT_OF_LOTTO_AT_LEAST_ONE = "적어도 하나 이상의 복권을 넣어야합니다."
-        fun of(lottos: List<Lotto>, winningLotto: WinningLotto): LottoResult =
-            LottoResult(
-                Rank.values().associateWith { rank ->
-                    lottos.count { rank == Rank.valueOf(winningLotto.getCountOfMatch(it), winningLotto.matchBonus(it)) }
-                },
-            )
     }
 }
