+++ src/main/kotlin/domain/WinningLotto.kt
@@ -7,9 +7,16 @@ class WinningLotto(val lotto: Lotto, val bonusNumber: LottoNumber) {
 
     constructor(numbers: IntArray, bonusNumber: Int) : this(Lotto(*numbers), LottoNumber(bonusNumber))
 
-    fun getCountOfMatch(anyLotto: Lotto): Int = anyLotto.count { lotto.contains(it) }
+    fun match(lottos: List<Lotto>): LottoResult =
+        LottoResult(
+            Rank.values().associateWith { rank ->
+                lottos.count { rank == Rank.valueOf(getCountOfMatch(it), matchBonus(it)) }
+            },
+        )
 
-    fun matchBonus(anyLotto: Lotto): Boolean = anyLotto.contains(bonusNumber)
+    private fun getCountOfMatch(anyLotto: Lotto): Int = anyLotto.count { lotto.contains(it) }
+
+    private fun matchBonus(anyLotto: Lotto): Boolean = anyLotto.contains(bonusNumber)
 
     companion object {
         private const val ERROR_DUPLICATE_BONUS_NUMBER = "보너스 번호는 당첨 번호와 중복될 수 없습니다. \n잘못된 값 : %s, %s"
