+++ src/test/kotlin/domain/LottoResultTest.kt
@@ -6,28 +6,6 @@ import org.junit.jupiter.api.Test
 
 class LottoResultTest {
     @Test
-    fun `로또 결과를 정상적으로 저장할 수 있다`() {
-        val lottos = listOf(
-            Lotto(1, 2, 3, 4, 5, 6), // 1등
-            Lotto(2, 3, 4, 5, 6, 7), // 2등
-            Lotto(3, 4, 5, 6, 7, 8), // 4등
-        )
-        val winningLotto = WinningLotto(intArrayOf(1, 2, 3, 4, 5, 6), 7)
-        val result = LottoResult.of(lottos, winningLotto)
-        val expect = LottoResult(
-            mapOf(
-                Rank.FIRST to 1,
-                Rank.SECOND to 1,
-                Rank.THIRD to 0,
-                Rank.FOURTH to 1,
-                Rank.FIFTH to 0,
-                Rank.MISS to 0,
-            ),
-        )
-        assertEquals(result, expect)
-    }
-
-    @Test
     fun `로또 수익률을 반환한다`() {
         val lottoResult = LottoResult(
             mapOf(
